requirements.txt
requests 
python-dotenv
pandas
matplotlib
scikit-learn